made panic handling a middleware + default app now includes tracing, panic, and logging middlewares
